+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/BuildResult.java
@@ -7,5 +7,10 @@
 package com.offbytwo.jenkins.model;
 
 public enum BuildResult {
-    FAILURE, UNSTABLE, REBUILDING, BUILDING, ABORTED, SUCCESS, UNKNOWN, NOT_BUILT
+    FAILURE, UNSTABLE, REBUILDING, BUILDING, ABORTED, SUCCESS, UNKNOWN, NOT_BUILT, 
+    /**
+     * This will be the result of a job in cases where it has been 
+     * cancelled during the time in the queue. 
+     */
+    CANCELLED
 }